Extract Defaults+Combine
